FBISCC-59: add UI tests to build pipeline
