+++ .eslintrc
@@ -1,5 +1,5 @@
 extends:
-- "homeoffice/config/default"
+  - "homeoffice/config/default"
 
 rules:
   consistent-return: off
@@ -8,6 +8,9 @@ env:
   es6: true
   node: true
   mocha: true
+
+parserOptions:
+  ecmaVersion: 8
 
 overrides:
   - files:
@@ -18,6 +21,9 @@ overrides:
       chai: true
       sinon: true
       expect: true
+      browser: true
+      page: true
+      baseURL: true
   - files:
       "assets/js/*.js"
     env:
