+++ .gitignore
@@ -5,3 +5,4 @@ apps/*/translations/*
 .idea
 .DS_Store
 apps/.DS_Store
+.nyc_output
