+++ package.json
@@ -2,11 +2,18 @@
   "license": "MIT",
   "scripts": {
     "start": "node app.js",
+    "start:silent": "node app.js silent",
     "dev": "hof-build watch",
     "debug": "node --inspect app.js",
     "test": "npm run test:lint && npm run test:unit",
     "test:lint": "eslint .",
-    "test:unit": "mocha",
+    "test:unit": "nyc mocha \"test/unit/**/*.spec.js\" --require test/test-helpers.js",
+    "test:ui": "npm run test:ui:firefox && npm run test:ui:webkit && npm run test:ui:chromium",
+    "test:ui:browser": "mocha \"test/ui/**/*.spec.js\" --require test/test-helpers.js test/ui/ui-test-helpers.js",
+    "test:ui:chromium": "npm run test:ui:browser -- --browser-engine=chromium",
+    "test:ui:firefox": "npm run test:ui:browser -- --browser-engine=firefox",
+    "test:ui:webkit": "npm run test:ui:browser -- --browser-engine=webkit",
+    "test:ui:server": "start-server-and-test start:silent http://localhost:8080/landing?automation-test test:ui",
     "build": "hof-build",
     "postinstall": "npm run build"
   },
@@ -27,11 +34,14 @@
     "chai": "^4.2.0",
     "eslint": "^5.16.0",
     "eslint-config-homeoffice": "^2.1.0",
+    "minimist": "^1.2.5",
     "mocha": "^8.1.3",
     "nyc": "^15.1.0",
+    "playwright": "^1.6.1",
     "proxyquire": "^2.1.3",
     "sinon": "^9.2.0",
-    "sinon-chai": "^3.5.0"
+    "sinon-chai": "^3.5.0",
+    "start-server-and-test": "^1.11.5"
   },
   "name": "fbis-ccf",
   "description": "Future Border and Immigration System (FBIS) Customer Contact Form (CCF)",
